Move the evdev specific code out of main.cpp
